Add hashing of password
Password will now be hashed using password_hash(). As of writing,
this method uses bcrypt. A salt is automatically generated and
stored with the password (hash:password).
Database column also changed to 255 if length of algorithm result
expands beyond the 60 chars that is used now.
A cost should be benchmarked and used specific for the site. The
default of 10 is now used. This is to be done in the future,
altough 10 is a good baseline cost.